requirements.txt
fastapi
uvicorn
openai
python-dotenv
PyMuPDF
python-docx
sentence-transformers
faiss-cpu
